config for autoloading license (oss parts)
