+++ command/server/config_test.go
@@ -1,5 +1,3 @@
-// +build !enterprise
-
 package server
 
 import (
@@ -10,16 +8,8 @@ func TestLoadConfigFile(t *testing.T) {
 	testLoadConfigFile(t)
 }
 
-func TestLoadConfigFile_topLevel(t *testing.T) {
-	testLoadConfigFile_topLevel(t, nil)
-}
-
 func TestLoadConfigFile_json(t *testing.T) {
 	testLoadConfigFile_json(t)
-}
-
-func TestLoadConfigFile_json2(t *testing.T) {
-	testLoadConfigFile_json2(t, nil)
 }
 
 func TestLoadConfigFileIntegerAndBooleanValues(t *testing.T) {
@@ -44,10 +34,6 @@ func TestConfig_Sanitized(t *testing.T) {
 
 func TestParseListeners(t *testing.T) {
 	testParseListeners(t)
-}
-
-func TestParseEntropy(t *testing.T) {
-	testParseEntropy(t, true)
 }
 
 func TestConfigRaftRetryJoin(t *testing.T) {
