+++ command/server/test-fixtures/config.hcl
@@ -48,3 +48,4 @@ disable_sealwrap = true
 disable_printable_check = true
 enable_response_header_hostname = true
 enable_response_header_raft_node_id = true
+license_path = "/path/to/license"
